Allow Custom Loading State
The error state is provied as a property of the render method. The error
is either one of 'Network Error' for connectivity problems or 'SDK Authentication Error'
for Google Maps errors.

You can set a custom loading state conditionally if there is no
googleMaps loaded and if there is no network error.

If there is an SDK Authentication error, googleMaps will still be
loaded, but with an error which you can show in the googleMaps
conditional block.

Refer to README file.